Yazar, Yayınevi arama eklendi

+++ 20Haziran/Form1.cs
@@ -18,25 +18,44 @@ namespace _20Haziran
     public partial class Form1 : Form
     {
         BindingSource kitapBindingSource = new BindingSource();
+        KitaplarDAO kitaplarDAO = new KitaplarDAO();
+        YayýneviDAO yayýneviDAO = new YayýneviDAO();
+        YazarDAO yazarDAO = new YazarDAO();
         public Form1()
         {
             InitializeComponent();
+
         }
 
         private void button1_Click(object sender, EventArgs e)
         {
-            KitaplarDAO kitaplarDAO = new KitaplarDAO();
+
             kitapBindingSource.DataSource = kitaplarDAO.getAllKitaplar();
-            dataGridView1.DataSource = kitapBindingSource;
+            columns(kitapBindingSource);
+
         }
 
         private void button2_Click(object sender, EventArgs e)
         {
-            KitaplarDAO kitaplarDAO = new KitaplarDAO();
+            if (string.IsNullOrEmpty(textBox1.Text))
+            {
+                MessageBox.Show("Lütfen bir kitap ismi giriniz.");
+                return;
+            }
+
             kitapBindingSource.DataSource = kitaplarDAO.searchTitles(textBox1.Text);
-            dataGridView1.DataSource = kitapBindingSource;
+            columns(kitapBindingSource);
+
+
         }
 
+        private void columns(BindingSource kitapBindingSource)
+        {
+            dataGridView1.DataSource = kitapBindingSource;
+            dataGridView1.Columns["yazarId"].Visible = false;
+            dataGridView1.Columns["yayýneviId"].Visible = false;
+
+        }
         private void button3_Click(object sender, EventArgs e)
         {
 
@@ -59,7 +78,7 @@ namespace _20Haziran
             {
                 int rowClicked = dataGridView1.CurrentRow.Index;
 
-                KitaplarDAO kitaplarDAO = new KitaplarDAO();
+
                 Kitap kitap = kitaplarDAO.gosterKitap((int)dataGridView1.Rows[rowClicked].Cells[0].Value);
 
                 Form3 form3 = new Form3(kitap);
@@ -73,12 +92,20 @@ namespace _20Haziran
 
         private void sil_button_Click(object sender, EventArgs e)
         {
+            if (dataGridView1.CurrentRow == null)
+            {
+                MessageBox.Show("Lütfen bir satýr seçiniz.");
+                return;
+            }
             int rowClicked = dataGridView1.CurrentRow.Index;
+            int sayi = rowClicked;
             int kitapID = (int)dataGridView1.Rows[rowClicked].Cells[0].Value;
 
-            KitaplarDAO kitaplarDAO = new KitaplarDAO();
+
             int result = kitaplarDAO.deleteKitap(kitapID);
             MessageBox.Show(result + " kayýt silindi");
+
+
         }
 
         private void YazarEkle_Click(object sender, EventArgs e)
@@ -104,6 +131,67 @@ namespace _20Haziran
         {
             YayýneviSil yayýnevisil = new YayýneviSil();
             yayýnevisil.Show();
+        }
+
+        private void YayýneviGoster_Click(object sender, EventArgs e)
+        {
+            kitapBindingSource.DataSource = yayýneviDAO.getAllYayýnevleri();
+            dataGridViewYayýnevi.DataSource = kitapBindingSource;
+        }
+
+        private void YayýneviAra_Click(object sender, EventArgs e)
+        {
+            if (string.IsNullOrEmpty(textBoxYayýnevi.Text))
+            {
+                MessageBox.Show("Lütfen bir yayýnevi ismi giriniz.");
+                return;
+            }
+
+            kitapBindingSource.DataSource = yayýneviDAO.searchTitles(textBoxYayýnevi.Text);
+            dataGridViewYayýnevi.DataSource = kitapBindingSource;
+
+        }
+
+        private void YazarGoster_Click(object sender, EventArgs e)
+        {
+            kitapBindingSource.DataSource = yazarDAO.getAllYazarlar();
+            dataGridViewYazar.DataSource = kitapBindingSource;
+        }
+
+        private void YazarAra_Click(object sender, EventArgs e)
+        {
+            if (string.IsNullOrEmpty(textBoxYazar.Text))
+            {
+                MessageBox.Show("Lütfen bir yazar ismi giriniz.");
+                return;
+            }
+
+            kitapBindingSource.DataSource = yazarDAO.searchTitles(textBoxYazar.Text);
+            dataGridViewYazar.DataSource = kitapBindingSource;
+        }
+
+        private void YazarAdýAra_Click(object sender, EventArgs e)
+        {
+            if (string.IsNullOrEmpty(textBoxYazarAdý.Text))
+            {
+                MessageBox.Show("Lütfen bir yazar ismi giriniz.");
+                return;
+            }
+
+            kitapBindingSource.DataSource = kitaplarDAO.searchAuthors(textBoxYazarAdý.Text);
+            columns(kitapBindingSource);
+        }
+
+        private void YayýneviAd_Click(object sender, EventArgs e)
+        {
+            if (string.IsNullOrEmpty(textBoxYayýneviAdý.Text))
+            {
+                MessageBox.Show("Lütfen bir yayýnevi ismi giriniz.");
+                return;
+            }
+
+            kitapBindingSource.DataSource = kitaplarDAO.searchPublishers(textBoxYayýneviAdý.Text);
+            columns(kitapBindingSource);
         }
     }
 }
